Seller/python/requirements.txt
flask
flask-cors
pymongo
scikit-learn
pandas
numpy
python-dotenv
joblib
